+++ packages/commons/ir-utils/src/examples/v1/generateTypeDeclarationExample.ts
@@ -171,6 +171,47 @@ export function generateTypeDeclarationExample({
                 }
             }
 
+            const baseProperties:ExampleObjectProperty[] = typeDeclaration.shape.baseProperties.map((property) => {
+                const propertyExample = generateTypeReferenceExample({
+                    fieldName: property.name.wireValue,
+                    typeReference: property.valueType,
+                    typeDeclarations,
+                    currentDepth: currentDepth + 1,
+                    maxDepth,
+                    skipOptionalProperties
+                });
+                if(propertyExample.type === "failure") {
+                    throw new Error(`Failed to generate example for union base property ${property.name.wireValue}`);
+                }
+                const { example  } = propertyExample;
+                return {
+                    name: property.name,
+                    originalTypeDeclaration: typeDeclaration.name,
+                    value: example,
+                    propertyAccess: property.propertyAccess,
+                };
+            });
+            const extendProperties: ExampleObjectProperty[] = typeDeclaration.shape.extends.flatMap((extendedTypeReference) => {
+                const extendedTypeDeclaration = typeDeclarations[extendedTypeReference.typeId];
+                if (extendedTypeDeclaration == null) {
+                    throw new Error(`Failed to find extended type declaration with id ${extendedTypeReference.typeId}`);
+                }
+                const extendedExample = generateTypeDeclarationExample({
+                    fieldName,
+                    typeDeclaration: extendedTypeDeclaration,
+                    typeDeclarations,
+                    currentDepth: currentDepth + 1,
+                    maxDepth,
+                    skipOptionalProperties
+                });
+                if (extendedExample == null || extendedExample.type === "failure") {
+                    throw new Error(`Failed to generate example for extended type declaration ${extendedTypeDeclaration.name.typeId}`);
+                }
+                if (extendedExample.example.type !== "object") {
+                    throw new Error(`Extended type declaration ${extendedTypeDeclaration.name.typeId} is not an object`);
+                }
+                return extendedExample.example.properties;
+            });
             for (const variant of typeDeclaration.shape.types) {
                 const variantExample = variant.shape._visit<ExampleGenerationResult<ExampleTypeShape>>({
                     noProperties: () => {
@@ -181,7 +222,9 @@ export function generateTypeDeclarationExample({
                                 singleUnionType: {
                                     wireDiscriminantValue: variant.discriminantValue,
                                     shape: ExampleSingleUnionTypeProperties.noProperties()
-                                }
+                                },
+                                baseProperties,
+                                extendProperties
                             }),
                             jsonExample: {
                                 [discriminant.wireValue]: variant.discriminantValue.wireValue,
@@ -228,7 +271,9 @@ export function generateTypeDeclarationExample({
                                                 ? example
                                                 : { properties: [], extraProperties: undefined }
                                     })
-                                }
+                                },
+                                baseProperties,
+                                extendProperties
                             }),
                             jsonExample: {
                                 [discriminant.wireValue]: variant.discriminantValue.wireValue,
@@ -257,7 +302,9 @@ export function generateTypeDeclarationExample({
                                 singleUnionType: {
                                     wireDiscriminantValue: variant.discriminantValue,
                                     shape: ExampleSingleUnionTypeProperties.singleProperty(example)
-                                }
+                                },
+                                baseProperties,
+                                extendProperties
                             }),
                             jsonExample: {
                                 [discriminant.wireValue]: variant.discriminantValue.wireValue,